+++ src/utils/browser.js
@@ -1,17 +1,26 @@
-import serialize from 'serialize-to-js';
 import { getUserAgentRegex } from 'browserslist-useragent-regexp';
 
 function getSupportedBrowserDetector(ignore) {
   if (ignore) {
     return JSON.stringify({ regex: '' });
   } else {
-    return serialize({
-      regex: getUserAgentRegex({
+    return JSON.stringify({
+      regex: 'value'
+    }).replace(
+      '"value"',
+      `new RegExp(${getUserAgentRegex({
         allowHigherVersions: true,
         allowZeroSubversions: true
-      })
-    });
+      })})`
+    );
   }
 }
 
-export { getSupportedBrowserDetector };
+function getCrossorigin(crossorigin, defaultCrossOrigin) {
+  crossorigin = crossorigin === true ? 'anonymous' : crossorigin;
+  crossorigin =
+    crossorigin === undefined ? defaultCrossOrigin || 'anonymous' : crossorigin;
+  return crossorigin;
+}
+
+export { getSupportedBrowserDetector, getCrossorigin };