+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Project Changelog
 
+## [3.0.11-next.3](https://github.com/basics/nuxt-booster/compare/v3.0.11-next.2...v3.0.11-next.3) (2024-04-30)
+
+
+### Bug Fixes
+
+* **build:** clean rollup ([c62bc10](https://github.com/basics/nuxt-booster/commit/c62bc10c7e1a33d1b0ce093870d91f69722683c9))
+
+## [3.0.11-next.2](https://github.com/basics/nuxt-booster/compare/v3.0.11-next.1...v3.0.11-next.2) (2024-04-30)
+
+
+### Bug Fixes
+
+* **blob:** fix import test ([1e226b8](https://github.com/basics/nuxt-booster/commit/1e226b80e55af83052a6582e769460d1317a6c33))
+
+## [3.0.11-next.1](https://github.com/basics/nuxt-booster/compare/v3.0.10...v3.0.11-next.1) (2024-04-30)
+
+
+### Bug Fixes
+
+* **composables:** fix rollup warnings ([d8796c6](https://github.com/basics/nuxt-booster/commit/d8796c667353d065611c6d1ae66f650cad8628c2))
+* **module:** remove `serialize-to-js` ([396afde](https://github.com/basics/nuxt-booster/commit/396afdec2655fca980bfa4ee60c6d7e9b2effde1))
+
 ## [3.0.10](https://github.com/basics/nuxt-booster/compare/v3.0.9...v3.0.10) (2024-04-29)